+++ libs/sdk-logux-designer/src/reference/reference.action.ts
@@ -5,11 +5,19 @@ import type { DesignerAction } from '@/types';
 
 const typeFactory = protocol.typeFactory('reference');
 
+interface ReferenceData {
+  references: Reference[];
+  referenceResources: ReferenceResource[];
+}
+
 export interface Replace extends DesignerAction {
-  data: {
-    references: Reference[];
-    referenceResources: ReferenceResource[];
-  };
+  data: ReferenceData;
 }
 
 export const Replace = protocol.createAction<Replace>(typeFactory('replace'));
+
+export interface AddMany extends DesignerAction {
+  data: ReferenceData;
+}
+
+export const AddMany = protocol.createAction<AddMany>(typeFactory('add-many'));