+++ apps/realtime/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.395.1](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime@1.395.0...@voiceflow/realtime@1.395.1) (2024-06-14)
+
+### Bug Fixes
+
+* make sure pathOrder has a value (COR-2270) ([#8306](https://github.com/voiceflow/creator-app/issues/8306)) ([f3ba187](https://github.com/voiceflow/creator-app/commit/f3ba1870d8a95a14d6659834d48f5c1c17724e55))
+
+# [1.395.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime@1.394.0...@voiceflow/realtime@1.395.0) (2024-06-14)
+
+### Features
+
+* reference add block action handler (COR-000) ([#8267](https://github.com/voiceflow/creator-app/issues/8267)) ([7e1312d](https://github.com/voiceflow/creator-app/commit/7e1312d3c656b6e3e89208b052bef6089dc2b6ff))
+
 # [1.394.0](https://github.com/voiceflow/creator-app/compare/@voiceflow/realtime@1.393.0...@voiceflow/realtime@1.394.0) (2024-06-14)
 
 ### Features